split OcdId class out into its own file
no reason why this should live within `combine_sources.rb`
